django-book-shop/requirements.txt
asgiref==3.3.4
Django==3.2.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
gunicorn
whitenoise
dj-database-url
python-dotenv
